backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.4.2

# Seismic Data Analysis Libraries
numpy==1.24.3
scipy==1.11.3
matplotlib==3.7.2
plotly==5.17.0
pandas==2.1.1
scikit-image==0.21.0
h5py==3.9.0
obspy==1.4.0
segyio==1.9.11

# 3D Visualization and Processing
vtk==9.3.0
pyvista==0.42.3
mayavi==4.8.1
pillow==10.0.1

# Machine Learning for Advanced Processing
scikit-learn==1.3.0
tensorflow==2.13.0
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.1.0

# Time Series Analysis
statsmodels==0.14.0
pmdarima==2.0.4

# Data Processing for Reservoir Analysis
openpyxl==3.1.2
xlrd==2.0.1

# Additional Utilities
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
